refactor: modularize match logic and UI components
